requirements.txt
torch>=2.0
torchvision>=0.15
streamlit>=1.30
numpy
pillow
scikit-learn
tqdm
